Allow CLI opts to override auth token and endpoint
Previously, both --os-auth-token and --os-image-url had to be
provided in order for either of them to be used in any API requests.
This breaks the tie between them and allows a user to override
either the auth token or the endpoint returned by Keystone independently
of one another.

Fixes bug 1029586

Change-Id: I8b81be723286c546d9cbd97c8b7d7aa89c03b2d4